pyproject.toml
[project]
name = "electro-retail"
version = "0.1.0"
description = ""
authors = [
    {name = "Ksenia Veremyeва",email = "ver.ksenia.16@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1,<2.0.0",
    "django>=6.0.1,<7.0.0",
    "django-filter>=25.2,<26.0",
    "psycopg2-binary>=2.9.11,<3.0.0",
    "djangorestframework>=3.16.1,<4.0.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "flake8 (>=7.3.0,<8.0.0)"
]
